requirements.txt
marshmallow~=3.21.1
pymongo~=4.6.2
flask~=2.2.2
python-dotenv~=1.0.1
werkzeug~=2.2.2
flask_pymongo
flask_jwt_extended
